more test fixes.
