+++ src/Dogger/Infrastructure/Configuration/ConfigurationFactory.cs
