+++ src/Dogger/Infrastructure/Configuration/IOnPremisesManifest.cs
