+++ src/Dogger/Infrastructure/Configuration/OnPremisesManifest.cs
